RAG/requirements.txt
streamlit
openai
PyMuPDF
numpy
python-dotenv
torch